# protocol/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "canix402 API",
    "version": "1.0.0",
    "description": "x402-gated Algorand DeFi data and walletless transaction API, including Haystack swap quotes, prerequisite opt-ins, and paid atomic swap-group generation. Numeric yield/USD fields (apy, apr, tvlUsd) are returned with a standard precision of 6 decimal places, extended up to 12 decimal places only when needed to represent small non-zero values. Asset decimals are always sourced on-chain from algod.",
    "contact": {
      "name": "Neon Forge Ltd",
      "email": "kieran@neonforge.ltd",
      "url": "https://canix402.compx.io"
    },
    "x-guidance": "Use the gateway base URL in servers[0]. Fetch /discovery and /openapi.json before integrating. Prefer the remote canix402 MCP server at https://canix402-mcp.compx.io/mcp (streamable-http, walletless pass-through) for agent tool access. Paid routes return HTTP 402 with PAYMENT-REQUIRED; sign a USDC ASA transfer client-side and retry with PAYMENT-SIGNATURE. Haystack swaps follow quote -> opt-in if needed -> fresh quote -> paid transaction generation -> local signing/submission. Human and LLM docs: https://canix402.compx.io/llms.txt",
    "x-precision": {
      "fields": [
        "apy",
        "apr",
        "tvlUsd"
      ],
      "defaultDecimalPlaces": 6,
      "maxDecimalPlaces": 12,
...
    }
  },
  "servers": [
    {
      "url": "https://canix402-api.compx.io"
    }
  ],
...
  "paths": {
...
    "/opportunities": {
      "get": {
        "summary": "Top 10 aggregated DeFi opportunities ranked by APY",
        "description": "Returns ranked Algorand DeFi yield opportunities across supported protocols including Tinyman, Pact, Folks Finance, CompX, Dork.fi, Myth Finance, and Haystack. Use when an agent needs to compare APY/APR, TVL, asset pairs, opportunity type, protocol, source freshness, and caveats before presenting or ranking yield options. This endpoint provides normalized market data only; it does not build or submit transactions.",
        "tags": [
          "defi",
          "opportunities"
        ],
        "parameters": [
          {
            "name": "protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "description": "Bypass Redis and refetch adapters; still writes a fresh cache entry. Default false.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunitiesListResponse"
                },
                "examples": {
                  "default": {
                    "$ref": "#/components/examples/AggregateOpportunitiesSample"
                  }
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.01"
          }
        },
        "x-x402": {
          "protocolVersion": 2,
          "requiredHeaders": [
            "PAYMENT-REQUIRED",
            "PAYMENT-SIGNATURE",
            "PAYMENT-RESPONSE"
          ],
          "requirementTemplate": {
            "scheme": "exact",
            "network": "algorand-mainnet",
            "asset": "31566704",
            "payTo": "REPLACE_WITH_PAYTO_ADDRESS",
            "maxAmountRequired": "0.01"
          }
        }
      }
    },
    "/opportunities/search": {
      "get": {
        "summary": "Caller-filtered opportunities across supported platforms",
        "description": "Returns Algorand DeFi opportunities filtered by platform, opportunity type, APY range, TVL threshold, and optional ASA assetIds across supported sources. Use when an agent needs targeted discovery such as high-yield liquidity pools, lending markets, protocol-specific yield, minimum-liquidity opportunities, or yields for specific Algorand assets (assetIds=0 for ALGO). This endpoint provides normalized market data only; it does not build or submit transactions.",
        "tags": [
          "defi",
          "opportunities",
          "search"
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Comma-separated list of platforms."
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Comma-separated list of opportunity types."
            }
          },
          {
            "name": "minApy",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxApy",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minTvlUsd",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "assetIds",
            "in": "query",
            "description": "Comma-separated Algorand ASA ids (0 = ALGO). Keeps opportunities whose assetIds intersect any requested id. Rows without assetIds do not match when this filter is set.",
            "schema": {
              "type": "string",
              "example": "0,31566704"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "description": "Bypass Redis and refetch adapters; still writes a fresh cache entry. Default false.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered opportunity list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunitiesListResponse"
                },
                "examples": {
                  "default": {
                    "$ref": "#/components/examples/FilteredOpportunitiesSample"
                  }
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.01"
          }
        },
        "x-x402": {
          "protocolVersion": 2,
          "requiredHeaders": [
            "PAYMENT-REQUIRED",
            "PAYMENT-SIGNATURE",
            "PAYMENT-RESPONSE"
          ],
          "requirementTemplate": {
            "scheme": "exact",
            "network": "algorand-mainnet",
            "asset": "31566704",
            "payTo": "REPLACE_WITH_PAYTO_ADDRESS",
            "maxAmountRequired": "0.01"
          }
        }
      }
    },
    "/opportunities/personalized": {
      "get": {
        "summary": "Top opportunities tuned to a wallet's held assets",
        "description": "Returns Algorand DeFi opportunities whose underlying assets match a supplied wallet's holdings, including opted-in ASAs with positive balance and native ALGO when held. Use when an agent needs wallet-aware yield ideas based on assets the account already owns, with normalized APY/APR, TVL, asset ids, source freshness, and caveats. This endpoint provides normalized market data only; it does not build or submit transactions.",
        "tags": [
          "defi",
          "opportunities",
          "personalized",
          "wallet"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Algorand account address to personalize opportunities for."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "description": "Bypass Redis and refetch adapters; still writes a fresh cache entry. Default false.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personalized opportunity list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizedOpportunitiesListResponse"
                },
                "examples": {
                  "default": {
                    "$ref": "#/components/examples/PersonalizedOpportunitiesSample"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidAddress"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.05"
          }
        },
        "x-x402": {
          "protocolVersion": 2,
          "requiredHeaders": [
            "PAYMENT-REQUIRED",
            "PAYMENT-SIGNATURE",
            "PAYMENT-RESPONSE"
          ],
          "requirementTemplate": {
            "scheme": "exact",
            "network": "algorand-mainnet",
            "asset": "31566704",
            "payTo": "REPLACE_WITH_PAYTO_ADDRESS",
            "maxAmountRequired": "0.05"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable error code.",
                "enum": [
                  "VALIDATION_ERROR",
                  "INTERNAL_ERROR",
                  "NOT_FOUND",
                  "MISSING_PAYMENT_SIGNATURE",
                  "MALFORMED_PAYMENT_SIGNATURE",
                  "EXPIRED_PAYMENT_PROOF"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "description": "Optional structured context for the error."
              }
            }
          }
        }
      },
...
      "OpportunitiesListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpportunityRecord"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/OpportunitiesListMeta"
          }
        }
      },
      "PersonalizedOpportunitiesListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonalizedOpportunityRecord"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PersonalizedOpportunitiesListMeta"
          }
        }
      },
...
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. The client must retry with a signed PAYMENT-SIGNATURE payload.",
        "headers": {
          "PAYMENT-REQUIRED": {
            "description": "Base64-encoded JSON x402 payment requirements the client must satisfy before retrying.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "missingPayment": {
                "$ref": "#/components/examples/MissingPaymentSignatureError"
              },
              "malformedPayment": {
                "$ref": "#/components/examples/MalformedPaymentSignatureError"
              },
              "expiredProof": {
                "$ref": "#/components/examples/ExpiredPaymentProofError"
              }
            }
          }
        }
      },
      "InvalidAddress": {
        "description": "Request validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "invalidAddress": {
                "$ref": "#/components/examples/InvalidAddressError"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "internalError": {
                "$ref": "#/components/examples/InternalError"
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "AggregateOpportunitiesSample": {
        "summary": "Top aggregated opportunities",
        "value": {
          "data": [
            {
              "protocol": "tinyman",
              "opportunityType": "lp",
              "opportunityId": "tinyman:pool:1002541853",
              "assetPair": "ALGO/USDC",
              "assetIds": [
                0,
                31566704
              ],
              "apr": 10.512,
              "apy": 12.5,
              "yieldBasis": "apy",
              "tvlUsd": 2450000.5,
              "sourceTimestamp": "2026-07-06T09:00:00.000Z",
              "fetchedAt": "2026-07-06T09:00:00.000Z",
              "notes": "sourceTimestamp reflects fetch time; upstream does not expose a per-row update time.",
              "executionReady": true,
              "executionShapes": [
                {
                  "shapeKey": "mainnet:tinyman:v2:addLiquidity:flexible",
                  "protocol": "tinyman",
                  "protocolVersion": "v2",
                  "action": "addLiquidity",
                  "variant": "flexible",
                  "title": "Tinyman v2 flexible add liquidity",
                  "summary": "Adds two-sided liquidity to an existing Tinyman AMM v2 pool using flexible amounts. Generates the documented 3-transaction group (asset1 transfer, asset2 transfer/payment, add_liquidity app call) as unsigned transactions.",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress",
                    "assetAId",
                    "assetAAmount",
                    "assetBId",
                    "assetBAmount",
                    "maxSlippageBps"
                  ],
                  "requiredAssetIds": [
                    0,
                    31566704
                  ],
                  "inputHints": {
                    "assetAId": 0,
                    "assetBId": 31566704,
                    "poolId": "tinyman:pool:1002541853"
                  }
                },
                {
                  "shapeKey": "mainnet:tinyman:v2:addLiquidity:initial",
                  "protocol": "tinyman",
                  "protocolVersion": "v2",
                  "action": "addLiquidity",
                  "variant": "initial",
                  "title": "Tinyman v2 initial add liquidity",
                  "summary": "Adds the first liquidity to a bootstrapped but empty Tinyman AMM v2 pool. Generates the documented 3-transaction group (asset1 transfer, asset2 transfer/payment, add_initial_liquidity app call) as unsigned transactions.",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress",
                    "assetAId",
                    "assetAAmount",
                    "assetBId",
                    "assetBAmount",
                    "maxSlippageBps"
                  ],
                  "requiredAssetIds": [
                    0,
                    31566704
                  ],
                  "inputHints": {
                    "assetAId": 0,
                    "assetBId": 31566704,
                    "poolId": "tinyman:pool:1002541853"
                  }
                },
                {
                  "shapeKey": "mainnet:tinyman:v2:addLiquidity:singleAsset",
                  "protocol": "tinyman",
                  "protocolVersion": "v2",
                  "action": "addLiquidity",
                  "variant": "singleAsset",
                  "title": "Tinyman v2 single-asset add liquidity",
                  "summary": "Adds one-sided liquidity to an existing Tinyman AMM v2 pool. Tinyman performs an internal swap to balance the deposit and mints pool tokens. Generates the documented 2-transaction group (asset transfer/payment, add_liquidity app call) as unsigned transactions.",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress",
                    "assetAId",
                    "assetBId",
                    "depositAssetId",
                    "depositAmount",
                    "maxSlippageBps"
                  ],
                  "requiredAssetIds": [
                    0,
                    31566704
                  ],
                  "inputHints": {
                    "assetAId": 0,
                    "assetBId": 31566704,
                    "poolId": "tinyman:pool:1002541853"
                  }
                }
              ],
              "compatibleExitShapes": []
            },
            {
              "protocol": "reti",
              "opportunityType": "staking",
              "opportunityId": "reti-staking-12",
              "assetPair": "ALGO",
              "assetIds": [
                0
              ],
              "apy": 8.5,
              "apr": 8.5,
              "yieldBasis": "apr",
              "tvlUsd": 125000.5,
              "sourceTimestamp": "2026-07-23T10:00:00.000Z",
              "fetchedAt": "2026-07-23T10:00:00.000Z",
              "entryRequirements": {
                "minAmount": {
                  "assetId": 0,
                  "amount": "1000000000"
                },
                "eligibilityFullyCheckable": true
              },
              "capacity": {
                "stakerSlotsRemaining": 20,
                "algoRoomMicroAlgos": "50000000000",
                "acceptingStake": true
              },
              "notes": "Réti validator-level consensus staking; quote-time eligibility checks are authoritative.",
              "executionReady": true,
              "executionShapes": [
                {
                  "shapeKey": "mainnet:reti:v1:stake:algo",
                  "protocol": "reti",
                  "protocolVersion": "v1",
                  "action": "stake",
                  "variant": "algo",
                  "title": "Réti ALGO stake to validator",
                  "summary": "Stakes ALGO to a Réti validator via ValidatorRegistry.addStake. The registry allocates into one of the validator's pools. Builds gas ×2, ALGO payment to the registry, addStake, and optional reward-token opt-in. Enforces minEntryStake, capacity, and entry gates at quote time.",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress",
                    "validatorId",
                    "amount"
                  ],
                  "requiredAssetIds": [
                    0
                  ],
                  "inputHints": {
                    "validatorId": 12,
                    "assetId": 0,
                    "depositAssetId": 0
                  }
                }
              ],
              "compatibleExitShapes": [
                {
                  "shapeKey": "mainnet:reti:v1:unstake:algo",
                  "protocol": "reti",
                  "protocolVersion": "v1",
                  "action": "unstake",
                  "variant": "algo",
                  "title": "Réti ALGO unstake from pool",
                  "summary": "Removes ALGO stake from a Réti staking pool via StakingPool.removeStake. Builds gas ×2, removeStake, and optional reward-token opt-in. Partial unstake must leave at least minEntryStake (or unstake the full balance).",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress",
                    "validatorId",
                    "poolAppId",
                    "amount"
                  ],
                  "requiredAssetIds": [
                    0
                  ],
                  "inputHints": {
                    "validatorId": 12,
                    "assetId": 0,
                    "depositAssetId": 0
                  }
                }
              ]
            },
            {
              "protocol": "folks-finance",
              "opportunityType": "lending",
              "opportunityId": "folks-lending-971254667",
              "assetPair": "USDC",
              "assetIds": [
                31566704
              ],
              "apy": 6.06,
              "yieldBasis": "apy",
              "tvlUsd": 29846.609471,
              "sourceTimestamp": "2026-07-06T08:55:00.000Z",
              "fetchedAt": "2026-07-06T09:00:00.000Z",
              "executionReady": true,
              "executionShapes": [
                {
                  "shapeKey": "mainnet:folks-finance:v2:setup:depositEscrow",
                  "protocol": "folks-finance",
                  "protocolVersion": "v2",
                  "action": "setup",
                  "variant": "depositEscrow",
                  "title": "Folks Finance v2 add deposit escrow",
                  "summary": "Creates and funds a new Folks Finance deposit escrow for the user. Returns a 3-transaction group that must be signed by both the user and the generated escrow account.",
                  "order": 0,
                  "requiredInputs": [
                    "userAddress"
                  ],
                  "requiredAssetIds": [
                    31566704
                  ],
                  "inputHints": {
                    "poolAppId": 971254667,
                    "assetId": 31566704
                  }
                },
                {
                  "shapeKey": "mainnet:folks-finance:v2:setup:optEscrowAsset",
                  "protocol": "folks-finance",
                  "protocolVersion": "v2",
                  "action": "setup",
                  "variant": "optEscrowAsset",
                  "title": "Folks Finance v2 opt deposit escrow into pool fAsset",
                  "summary": "Funds and opts an existing Folks Finance deposit escrow into a pool fAsset so it can receive deposits.",
                  "order": 1,
                  "prerequisiteShapeKeys": [
                    "mainnet:folks-finance:v2:setup:depositEscrow"
                  ],
                  "requiredInputs": [
                    "userAddress",
                    "escrowAddress"
                  ],
                  "requiredAssetIds": [
                    31566704
                  ],
                  "inputHints": {
                    "poolAppId": 971254667,
                    "assetId": 31566704
                  }
                },
                {
                  "shapeKey": "mainnet:folks-finance:v2:deposit:escrow",
                  "protocol": "folks-finance",
                  "protocolVersion": "v2",
                  "action": "deposit",
                  "variant": "escrow",
                  "title": "Folks Finance v2 escrow deposit",
                  "summary": "Deposits an underlying asset into a Folks Finance lending pool via a deposit escrow. fAssets accrue in the escrow rather than the user's wallet. Generates the SDK deposit group (asset transfer + deposit app call), optionally prefixed with OpUp.",
                  "order": 2,
                  "prerequisiteShapeKeys": [
                    "mainnet:folks-finance:v2:setup:optEscrowAsset"
                  ],
                  "requiredInputs": [
                    "userAddress",
                    "assetAmount"
                  ],
                  "requiredAssetIds": [
                    31566704
                  ],
                  "inputHints": {
                    "poolAppId": 971254667,
                    "assetId": 31566704
                  }
                }
              ],
              "compatibleExitShapes": []
            }
          ],
          "meta": {
            "limit": 10,
            "offset": 0,
            "includeInactive": false,
            "paymentRequired": true,
            "cacheEnabled": false,
            "cacheHit": false,
            "cachedAt": null,
            "cacheAgeMs": null,
            "cacheTtlSec": 180
          }
        }
      },
...
      "MissingPaymentSignatureError": {
        "summary": "Paid route called without PAYMENT-SIGNATURE",
        "value": {
          "error": {
            "code": "MISSING_PAYMENT_SIGNATURE",
            "message": "The paid endpoint was called without PAYMENT-SIGNATURE."
          }
        }
      },
...
    }
  }
}